Include 'Low' blueprints in release status board
n the Icehosue cycle we used "Low" priority to fly below release
radar, so we excluded 'Low' priority blueprints from the release
status board, unless it was 'Implemented'.

Starting with Juno and with most projects now using the targeted
milestone list as a communication tool, any targeted blueprint
(as long as it's prioritized) should show up on the release status
board.

Change-Id: Ideff54642f89f1decbc49a14bc44604cf5bdd35d